Update configs for production (#128)
- Vite Config Is Now More Configurable 
- Https is no longer assumed out of the box, you need to specify it.
This is needed for Azure Container App or Kubernetes, as the SSL stuff
is handled not through the app, but rather through their infrastructure.
Still need it for local development, so docker compose was modified to
allow that
- The API proxy redirect also needs to be configurable, as that will
change on the production environment
- Client should now be using Dockerfile For deploying the client as a
front end container, I added a docker file for it. Updated the docker
compose to refer to said docker file. I also added in a copy of the
docker ignore file as well
- Updated Front End Deployment Github Action Tweaked the auto generated
file for the deployment for the azure app container service, hopefully
it should now deploy
- The front end should now have a health check in place