Enable local only Cypress Tests + Vitest setup (#248)
Originally was going to replace cypress with Vitest or playwright, but
that turned out to be harder then I care to admit.

Found out I can disable the cypress cloud integration, and decided to
keep around vitest for specific class testing

With that disabled, added new github actions that will allow me to see
the test state on the PR, and that comes with some nice integrations

Went through and fixed issues with the cypress tests. One of the biggest
problems I had was that I used the actual router instance in all
components, as opposed to relying on the injected one from main.ts,
which caused issues with the unit tests

fixed the warning regarding the faveicon and ifIHadOne images